plant_lib_be/requirements.txt
# --- Core FastAPI ---
fastapi==0.115.2
uvicorn[standard]==0.30.6

# --- Database ---
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary>=2.9,<3.0

# --- Pydantic & Config ---
pydantic==2.10.4
pydantic-settings==2.10.1

# --- Security / Auth ---
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# --- Utils ---
python-dotenv==1.0.1
email-validator==2.2.0
requests>=2.32.5,<3.0.0

# --- KG Pipeline (Neo4j + LLM) ---
google-generativeai>=0.7.0,<0.8.0
langchain==0.3.27
langchain-community==0.3.31
langchain-google-genai==2.0.0
langchain-experimental==0.3.0
neo4j==5.20.0
sentence-transformers==4.0.0
torch==2.6.0
transformers==4.51.0
accelerate==0.34.2
Pillow==11.1.0
langdetect==1.0.9
pandas==2.3.0
tqdm==4.67.1
python-multipart==0.0.9
